Clarify machine implementation ownership and runtime strategies
