+++ .changeset/clear-machine-ownership.md
@@ -1,0 +1,7 @@
+---
+"@typeonce/effect-machine": patch
+---
+
+Clarify declarative guards in the machine documentation and add a checked example of guarded state construction.
+
+Improve internal typing and organization for machine definitions, execution strategies, snapshot validation, and Cluster contracts. Existing public APIs and machine behavior remain unchanged.
