+++ packages/effect-machine/test/internal/machine/processLifecycle.test.ts
@@ -2,6 +2,7 @@ import { assert, describe, it } from "@effect/vitest"
 import { Cause, Deferred, Effect, Exit, Fiber, Option, Ref, Stream } from "effect"
 import { Machine } from "../../../src/index.js"
 import * as MachineRuntime from "../../../src/internal/machine/runtime.js"
+import * as MachineRuntimeProtocol from "../../../src/internal/machine/runtimeProtocol.js"
 
 describe("machine process lifecycle", () => {
   it.effect("reuses a settled active startup without running an empty compiled drain", () =>
@@ -222,7 +223,7 @@ describe("machine process lifecycle", () => {
 
   it.effect("provides empty child operations for childless process logic", () =>
     Effect.gen(function*() {
-      const processScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const processScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const ref = yield* MachineRuntime.startProcess({
         execution: { _tag: "Childless" },
         initial: (scope) => Deferred.succeed(processScope, scope).pipe(Effect.as(0)),
@@ -506,9 +507,9 @@ describe("machine process lifecycle", () => {
 
   it.effect("interrupts the worker before publishing an externally requested failure", () =>
     Effect.gen(function*() {
-      const runtime = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const runtime = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const cleanupCount = yield* Ref.make(0)
-      const logic: MachineRuntime.ProcessLogic<number, never, string> = {
+      const logic: MachineRuntimeProtocol.ProcessLogic<number, never, string> = {
         initial: (scope) => Deferred.succeed(runtime, scope).pipe(Effect.as(1)),
         run: () =>
           Effect.never.pipe(
@@ -690,7 +691,7 @@ describe("machine process lifecycle", () => {
         Array.from({ length: 50 }),
         () =>
           Effect.gen(function*() {
-            const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+            const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
             const parent = yield* MachineRuntime.startProcess({
               initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
               run: () => Effect.never
@@ -723,7 +724,7 @@ describe("machine process lifecycle", () => {
 
   it.effect("publishes every named child replacement in order", () =>
     Effect.gen(function*() {
-      const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const parent = yield* MachineRuntime.startProcess({
         initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
         run: () => Effect.never
@@ -756,7 +757,7 @@ describe("machine process lifecycle", () => {
 
   it.effect("preserves registry-wide childChanges emission ticks", () =>
     Effect.gen(function*() {
-      const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const parent = yield* MachineRuntime.startProcess({
         initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
         run: () => Effect.never
@@ -784,7 +785,7 @@ describe("machine process lifecycle", () => {
 
   it.effect("buffers ordered childChanges while a subscriber is stalled", () =>
     Effect.gen(function*() {
-      const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const parent = yield* MachineRuntime.startProcess({
         initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
         run: () => Effect.never
@@ -868,7 +869,7 @@ describe("machine process lifecycle", () => {
 
   it.effect("keeps child interruption parallel with scoped resource finalization", () =>
     Effect.gen(function*() {
-      const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const childInterrupted = yield* Deferred.make<void>()
       const resourceFinalized = yield* Deferred.make<void>()
       const release = yield* Deferred.make<void>()
@@ -909,11 +910,11 @@ describe("machine process lifecycle", () => {
 
   it.effect("does not orphan a child when parent stop races child initialization", () =>
     Effect.gen(function*() {
-      const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const childInitializing = yield* Deferred.make<void>()
       const releaseChild = yield* Deferred.make<void>()
       const resourceCleanup = yield* Ref.make(0)
-      const parentLogic: MachineRuntime.ProcessLogic<undefined, never> = {
+      const parentLogic: MachineRuntimeProtocol.ProcessLogic<undefined, never> = {
         initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
         run: () => Effect.never
       }
@@ -954,9 +955,9 @@ describe("machine process lifecycle", () => {
         Array.from({ length: 50 }),
         () =>
           Effect.gen(function*() {
-            const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+            const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
             const race = yield* Deferred.make<void>()
-            const parentLogic: MachineRuntime.ProcessLogic<undefined, never> = {
+            const parentLogic: MachineRuntimeProtocol.ProcessLogic<undefined, never> = {
               initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
               run: () => Effect.never
             }
@@ -999,16 +1000,17 @@ describe("machine process lifecycle", () => {
 
   it.effect("delivers done, failure, and stopped child outcomes exactly once", () =>
     Effect.gen(function*() {
-      const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const outcomes = yield* Ref.make<ReadonlyArray<string>>([])
-      const parentLogic: MachineRuntime.ProcessLogic<undefined, never> = {
+      const parentLogic: MachineRuntimeProtocol.ProcessLogic<undefined, never> = {
         initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
         run: () => Effect.never
       }
       const parent = yield* MachineRuntime.startProcess(parentLogic)
       const scope = yield* Deferred.await(parentScope)
-      const recordOutcome = (id: string) => (outcome: MachineRuntime.RuntimeOutcome<unknown, unknown, unknown>) =>
-        Ref.update(outcomes, (current) => [...current, `${id}:${outcome._tag}`])
+      const recordOutcome =
+        (id: string) => (outcome: MachineRuntimeProtocol.RuntimeOutcome<unknown, unknown, unknown>) =>
+          Ref.update(outcomes, (current) => [...current, `${id}:${outcome._tag}`])
 
       const done = yield* scope.spawn(
         Machine.logic({ initial: 0, run: () => Effect.succeed("output") }),
@@ -1038,8 +1040,8 @@ describe("machine process lifecycle", () => {
 
   it.effect("isolates child terminalization from outcome callback defects", () =>
     Effect.gen(function*() {
-      const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
-      const parentLogic: MachineRuntime.ProcessLogic<undefined, never> = {
+      const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
+      const parentLogic: MachineRuntimeProtocol.ProcessLogic<undefined, never> = {
         initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
         run: () => Effect.never
       }
@@ -1061,7 +1063,7 @@ describe("machine process lifecycle", () => {
 
   it.effect("delivers committed active child snapshots directly and in order", () =>
     Effect.gen(function*() {
-      const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const snapshots = yield* Ref.make<ReadonlyArray<number>>([])
       const parent = yield* MachineRuntime.startProcess({
         initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
@@ -1078,7 +1080,7 @@ describe("machine process lifecycle", () => {
         }),
         {
           id: "child",
-          [MachineRuntime.activeSnapshotObserver]: (snapshot) =>
+          [MachineRuntimeProtocol.activeSnapshotObserver]: (snapshot) =>
             Ref.update(snapshots, (current) => [...current, snapshot.state])
         }
       )
@@ -1090,7 +1092,7 @@ describe("machine process lifecycle", () => {
 
   it.effect("isolates active child state updates from snapshot callback defects", () =>
     Effect.gen(function*() {
-      const parentScope = yield* Deferred.make<MachineRuntime.ProcessScope<never>>()
+      const parentScope = yield* Deferred.make<MachineRuntimeProtocol.ProcessScope<never>>()
       const parent = yield* MachineRuntime.startProcess({
         initial: (scope) => Deferred.succeed(parentScope, scope).pipe(Effect.as(undefined)),
         run: () => Effect.never
@@ -1100,7 +1102,7 @@ describe("machine process lifecycle", () => {
           initial: 0,
           run: ({ setState }) => setState(1).pipe(Effect.as("output"))
         }),
-        { id: "child", [MachineRuntime.activeSnapshotObserver]: () => Effect.die("callback defect") }
+        { id: "child", [MachineRuntimeProtocol.activeSnapshotObserver]: () => Effect.die("callback defect") }
       )
 
       assert.strictEqual(yield* child.join, "output")
